refactor(#59): study benefit 연관 작업
- study benefit과 관련되어 String으로 프론트에서 전달 받던 부분을 StudyBenefitRequestDto로 수정
- 테스트에서 메서드 변경
